BITMAG-242, BITMAG-259: Rework after https://sbforge.org/fisheye/cru/CR-BITMAG-34
git-svn-id: https://sbforge.org/svn/bitrepository/trunk@240 c5b8cdcd-6091-0410-b920-e50e972e9166

+++ bitrepository-access-client/src/main/java/org/bitrepository/access/getfile/GetFileClientSettings.java
@@ -26,9 +26,12 @@ package org.bitrepository.access.getfile;
 
 import org.bitrepository.protocol.bitrepositorycollection.ClientSettings;
 
+/**
+ * Contains the GetFile specific settings for a client.
+ */
 public interface GetFileClientSettings extends ClientSettings {
     /**
      * Return the default timeout for waiting for a getFile request to finish. 
      */
-    int getGetFileDefaultTimeout();
+    long getGetFileDefaultTimeout();
 }
